Show local changes in list
When listing the status of the current directories in the workspace,
show whether or not each repository has local changes. This includes new
files (untracked changes) as well as changes to currently track files.
Once changes are committed, however, this will not show any information.
We should add that as a new feature.